requirements.txt
sqlalchemy
graphviz
pydot
google-genai
matplotlib
pandas
numpy
python-dotenv
sqlalchemy-schemadisplay
cairosvg
pillow
